pyproject.toml
[project]
name = "py-ai-1"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "devtools>=0.12.2",
    "logfire>=2.11.0",
    "pydantic-ai[examples,logfire]>=0.0.15",
    "python-dotenv>=1.0.1",
    "streamlit>=1.41.1",
]

[tool.logfire]
ignore_no_config = true
